requirements.txt
discord-rewrite
Pillow
PyNaCl
requests
beautifulsoup4
Ptable
PyNaCl
youtube_dl
yarl<1.2